Images with packaged linux headers
Summary:
For host machines without linux headers (e.g. Container optimized OS),
we package in a version of the headers that we can use for BPF.

This packaged headers are not used yet. That is coming in a separate diff.

Also, the packaged headers will be trimmed in a future diff as well.

Test Plan: None

Reviewers: zasgar, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D1219

GitOrigin-RevId: 2856e883c80246d06f0bd569906e923730b5202e