Add operator and protobuf nodes for "zip" operator.
Summary:
This operator is like a union across N inputs. The main distinction is
when it is applied to inputs which are ordered by time
(have a time_ column), it will ensure that the output
result is ordered by time_ as well. In the case that there
is no time_ column in one or more of the tables, it will
behave like a normal union.

In the current approach, there can be columns in input relation A
that are not present in input relation B and vice versa, but any columns
in A and B with the same name must also share the same type.

so zip on A [x: int, y: float] and B [y: float, z: string] is ok
and produces [x: int, y: float, z: string] as the output relation.
But [x: int, y:float] and [x: string] would return an error in this
approach.

Test Plan:
added a test which parses the zip from protobuf,
and computes output relations in both success and error cases.

Reviewers: michelle, philkuz, zasgar, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D1182

GitOrigin-RevId: 41945029ed3f8a2b3d11197eb6c4f264e0d87ef4