Add GRPC Source/Sink, RowBatch service protos
Summary:
This diff is the first step of supporting queries across multiple machines. It defines a GRPC service that will be used to send RowBatches between nodes. It also adds the proto / plan node definitions for the operators that will be using that service.

Next will be the exec nodes for GRPC source and Sink. After that will be the definition of the GRPC server that runs on each machine. It will receive row batches from various operators and funnel them to the correct (local) GRPC source.

Test Plan: Added some unit tests for proto parsing

Reviewers: michelle, philkuz, zasgar, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D1200

GitOrigin-RevId: 71e478643512c6815242e19f584e31dbc2fc576f