cleanup of Union (zip) node
Summary:
A few items to clean up zip node before I finish the exec node for it.

1) Rename to Union, which is more canonical
2) previously the column mapping was an array of output indexes. This was to support
the output relation being the union of all of the input relations. Aka a,b and b,c would
output a,b,c. This is more trouble than it's worth, especially with the lack of full support
for null values right now. As a result the specification changed to make more sense.
Now "column mapping" is an array of input indexes. so [4,1] for input 5 would be saying
to take column 4 of input 5 to produce the first output column of the union operator.
3) other cleanup

Test Plan: added a couple more tests

Reviewers: michelle, zasgar, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D1239

GitOrigin-RevId: 665484b1192ee19468bb5219a9d220df5a0aac6f