ExecNodeTester can have one ConsumeNext result in multiple row batches
Summary:
A single call to ConsumeNext in an exec node can result in multiple output row batches being passed to the child. This diff extends the ExecNodeTester to support that case.

Example usage:

In the case where input row batch A produces nothing, but then input row batch B comes in and output row batches 1 and 2 are produced:

tester.ConsumeNext(<A>).ConsumeNext(<A>).ExpectRowBatch(<1>).ExpectRowBatch(<2>)
would be the call.

Test Plan: existing should pass

Reviewers: michelle, zasgar, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D1264

GitOrigin-RevId: 7607e200e27c9685fbc1d58d834f07c3949a64fb