Utilities for finding or install linux headers (for BPF)
Summary:
As part of the linux headers pacakaging, these are utilities for finding
or installing linux headers into the container.

Decided to do this in C++ instead of a go wrapper, because they are really the domain of Stirling. They are required for stirling_wrapper as well, for example.

Test Plan:
Bunch of tests for the helper functions. I didn't create a test the top-level call (tested it manually instead),
because it's a bit tricker to test. Will take a TODO for that.

Reviewers: zasgar, yzhao, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D1222

GitOrigin-RevId: 5cd9c27b6770d2f353f72bee8ec51f3142473421