Kprobe cleaner functions and utility
Summary:
Even though we try to clean-up after ourselves,
there will be times when we leave dangling kprobes.

This is a utility to clean those probes up.

In a future diff, this will be incorporating into Stirling startup,
so we clean-up probes from any previous crash.

Test Plan: Added a test script that leaks probes, and then cleans them up, but it is only meant to be run manually for now.

Reviewers: zasgar, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D1268

GitOrigin-RevId: 5d84d7528d05a698a3c79f8a1a57805bdc8a14e7