Stirling: Signal handling
Summary:
To reduce the cases in which we leak kprobes, we need to handle signals.

This shows how it can be done on the stirling_wrapper process.

There is still a lot of future work left to future diffs, including at least:
1) Signal handling in the agent
2) Use of signal handling thread
3) Adapting the manual shell test for Bazel and Jenkins.

Test Plan: A test was added for PrepareToDie, and a manual shell test was added for the whole process.

Reviewers: zasgar, yzhao, #engineering

Reviewed By: yzhao, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D1242

GitOrigin-RevId: 2ee383444335cda80bf43e0ec890806c08299304