On AgentRegister, metadata service should add all k8s updates to agent's queue
Summary:
Before, we were sending k8s metadata updates in the AgentRegisterResponse.
Instead, we should send all of the metadata updates in the next Heartbeat.
There could potentially be a lot of k8s metadata updates that need to be queued up... And adding them in one at a time to etcd sounds painful. Instead, we add all the updates to the agent queue in a single batch transaction.

Test Plan: unit tests

Reviewers: zasgar, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D1288

GitOrigin-RevId: 886f598f614bad08b170c5aa06fffdc8ef738f06