Add nightly regression tests to jenkins
Summary:
We have noticed a few flappers and errors with repeated tests executions. This adds nightly tests
that will run the tests in a loop.

For some of our stress tests we might want to make the seeds configurable, because right now it only catches
chance bugs due to timing.

Test Plan: Jenkins

Reviewers: oazizi, #engineering

Reviewed By: oazizi, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D1324

GitOrigin-RevId: eb0e2c026c4999226ce815554e62700b893f4d1a