Compatibility with container-optimized OS (temporary workaroud)
Summary:
Container-optimized OS doesn't have /usr/src, only /usr,
so mount that instead.

Unfortunately kubernetes does not like it if the host directory of a mount doesn't exist.

And using DirectoryOrCreate doesn't work because the host machine is read-only.

So this is a temporary workaround, until we find a more robust solution.

***There will certainly be more diffs on this file.

Test Plan: I'm kind of testing in production here.

Reviewers: zasgar, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D1327

GitOrigin-RevId: 3e95b9cf3313df4264044633a6160d83a53a7722