Add end of window marker to row batches
Summary: When we want to add more windowed operations, we can either compute the window dynamically based on the timestamp or we can actually have an explicit end of window marker. For uneven windows (maybe based on session or something like that) then it is a lot more efficient to just window in one place and have everyone read those markers instead of computing it on the fly in each node. Aas a result this diff adding eow is the first step in supporting more windowed operations for flexible windows.

Test Plan: updated existing

Reviewers: michelle, philkuz, zasgar, #engineering

Reviewed By: michelle, #engineering

Subscribers: yzhao

Differential Revision: https://phab.corp.pixielabs.ai/D1316

GitOrigin-RevId: 5923cd2b230427c57ebec11ac45586b008dc7627