Moving common functions to BCCConnector
Summary:
The original intent was to keep BCCConnector as a base class,
with common functionalities of any BCC source connector that we build.

Unfortunately, we deviated from that over time.

This diff gets us back to the original intent, and refactors a bunch of
common functionalities into BCCConnector.

Test Plan: Existing tests

Reviewers: yzhao, zasgar, #engineering

Reviewed By: yzhao, zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D1190

GitOrigin-RevId: f090ef89d18790fb7406ec5c01acd382c99e801c