MetadataResolver Conversion rule
Summary:
The conversion rule takes a metadata resolver node and converts it into a Map node that has a physical
representation. This was a longer commit and would appreciate deeper feedback on this one.

Test Plan: Added new conversion rule tests to make sure that all the nodes are updated as expected.

Reviewers: michelle, zasgar, nserrino, #engineering

Reviewed By: michelle, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D1286

GitOrigin-RevId: 28a01b878576bd28840c6a19cf25dbf96e6d6534