Replace with EventType with TrafficDirection.
Summary:
This helps unifying the tracing code on the entry and return side of
the egress and ingress traffic.

That's done by reduce the need to pass in the event type, as the
direction is always inferable from the context, unlike the event type,
which has to from the highest-level probe functions.

Test Plan: Jenkins

Reviewers: chengruizhe, oazizi, zasgar, #engineering

Reviewed By: oazizi, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D1354

GitOrigin-RevId: 7267e167b4e2406c3c04cdbd51d0fbf3fffdc23d