Fixed segfault bug from indexing out of range, and added checks for future query and distribution additions
Summary: Added a check in string generator to ensure that the max and min output of the index generator falls within the range of the string vector.

Test Plan: Ran the entire benchmark under different optimizations, which I failed to do last time. Interested in learning a more automated approach for future additions.

Reviewers: #engineering, zasgar

Reviewed By: #engineering, zasgar

Subscribers: yzhao, zasgar

Differential Revision: https://phab.corp.pixielabs.ai/D1365

GitOrigin-RevId: 410e9ac9b6dc291809a30f5482929093e4cf1a07