Adding  MetadataFunctionFormatRule.
Summary:
MetadataFunctionFormatRule checks to make sure that any functions that have both metadata and another arg are formatted correctly.
This ishonestly why I rewrote hte analyzer of the compiler in the first place.

Test Plan: Added new tests in the rule_tests, will update analyzer later on when everything else is in place.

Reviewers: zasgar, michelle, nserrino, #engineering

Reviewed By: michelle, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D1260

GitOrigin-RevId: 1625e5c21c9c914d8f3a52f5804a66ead3f1266e