BPF modifications for older kernels (4.14)
Summary:
Our BPF code was not working on kernel 4.14.

This is because the BPF verifier:
1) Falsely believes that the bpf_probe_read size might be zero
2) And a bpf_probe_read of size zero is not allowed until kernel 4.15

Fix is to copy one extra byte, to make verifier happy.

Note that I've played around with many different styles,
to try to hint at the verifier that everything is okay,
but it's quite finicky.

The way the code is written is the one I've found that can keep the verifier happy.

Test Plan: Manually test on kernel 4.14.127+ (GKE).

Reviewers: yzhao, zasgar, #engineering

Reviewed By: yzhao, zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D1364

GitOrigin-RevId: 50f174d72eb4fb7a1d96d9d37f724ce910347803