Add RowBatchQueue to ExecState and EngineState
Summary:
Exec nodes need to have access to a queue to put row batches into when they need to send those row batches to a node elsewhere (like over GRPC). That queue is held by the exec state here. When carnot is initialized in the agent/compute node, this queue will be passed to it. This queue will be watched for row batches, and as they come in they will be sent to the appropriate destination over GRPC.

I wasn't sure about this being a shared_ptr. TableStore is a shared_ptr and this seems similar (the queue is owned by the enqueuer and the dequeuer) so I followed the pattern. However I am fine with making it a raw pointer as well.

Test Plan: This will get tested in two ways: on the enqueuing side (GRPC Sink exec node, coming soon) and on the dequeuing side (the thing that watches for new row batch messages and sends them to the appropriate destination, also coming soon).

Reviewers: michelle, zasgar, philkuz, #engineering

Reviewed By: michelle, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D1397

GitOrigin-RevId: b7cdf1820a48107fa471622429cc41005949acbf