Making columns track their own parents rather than using the parent operator.
Summary: I need to add support for multiple parents in operators, and this is part of shaving that yak. Removing this dependency makes it much more clear who the parent is when we start allowing multiple parents per op.

Test Plan: Tests pass, and since this doesn't add functionality this is sufficient.

Reviewers: zasgar, michelle, #engineering

Reviewed By: michelle, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D1403

GitOrigin-RevId: f074db1ac7416f6643e7bce5961fd49afac80042