MySQL Stitcher framework
Summary:
The framework of MySQL Stitcher with the stitch functions to show the general flow of stitching mysql packets.
Handlers are left empty and will be implemented in the next diff.
Removed previously duplicate MySQLEventType and fixed linkage in mysql_parser.
Add a seperate bazel cc_library for mysql.

Test Plan: existing tests

Reviewers: oazizi, yzhao, zasgar, #engineering

Reviewed By: oazizi, yzhao, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D1394

GitOrigin-RevId: 102ccc020c2c8a61be458ed986344e6a82898898