Add Union exec node
Summary:
This union has two cases, one where it merges the input row batches
by time and one where it just reorders the columns as specified (the unordered case).
A cursor is maintained across each parent stream to keep track of where we are in the merge. For performance reasons, in the ordered case, we buffer the data in the output row batch until it reaches a certain size. For the unordered case we just push the data as it comes in regardless of the size.

Test Plan: added test

Reviewers: michelle, zasgar, philkuz, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D1287

GitOrigin-RevId: fc63d0fb4d786bd3b31cb74a04f5db7709e684fe