Add QueuedRowBatch protobuf type for GRPC sink/sources
Summary: This diff is going to be a building block for the GRPC source/sink change. We will have a protobuf type that is queued up by carnot that represents a RowBatch that needs to be sent to another node for execution (in a multi-node query). This protobuf will be the element type of this queue. The GRPC sink node will enqueue these elements via the ExecState, and another function running parallel to Carnot will watch the queue for elements and ship them off via the GRPC service.

Test Plan: n/a

Reviewers: zasgar, michelle, philkuz, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D1309

GitOrigin-RevId: 79af095621daaf1de42ccfa150f47f7092578221