Consolidate InfoClassSchema and DataTableSchema
Summary:
There was a lot of overlap between InfoClassSchema and DataTableSchema,
and DataTableSchema is increasingly becoming our source of truth.

So kill InfoClassSchema.

With this change, there is some risk that we want to keep some dynamic state on the schema in the future, which would cause problems.
But we can cross that bridge if we ever come to it.

Probably the right thing to do in that cass is to add an auxiliary structure to InfoClassManager in that case.

Test Plan: Existing tests

Reviewers: zasgar, yzhao, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D1442

GitOrigin-RevId: 8eca3f3a32e1ae11d953baf37e010b41bb842d15