Stirling: Tabletized DataTableSchema
Summary:
This creates a tabletized DataTableSchema in Stirling.

This version still keeps the tabletized column in the tablet.
This causes redundant information, but is being done to simplify Carnot, so it does not need to understand tabletization.

This is in contrast to D1248, where the tablet value was not also repeated in the tablet (as it would be a redundant constant).

Test Plan:
Some coverage by stirling_test. Other coverage is harder to test, since DataTableSchema is mostly a constexpr structure.

Will add more tests once the tabletization key is exposed through the proto (future diff).

Reviewers: zasgar, philkuz, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D1440

GitOrigin-RevId: c4d6296e99448d848c2f04865c0a4211dfb522af