Stirling: Connection tracker benchmarking code
Summary:
This was a benchmark from a while ago to benchmark the container holding connection trackers.

I haven't bothered to clean it up. It's got too many ifdefs, and some other problems.

But it's not high enough priority to clean-up.

If there are too many nits on it--which is okay--I will prefer just not checking it in.

Test Plan: None - benchmark

Reviewers: zasgar, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D1461

GitOrigin-RevId: 41b40194c753eeaba8d3ad7d9fa34c3eb82c6ad9