PL-787: Adding support to specify multiple parents in Query language.
Summary: This is the first in a set of two commits to add support for multiple parents in the query language. The first step was to make sure that we can actually specify multiple parents using lambdas. The next is that we add the parent operator as part of the Column init instead of delaying the parent setting  in the rules which is too late when we get into the multiple-parent regime.

Test Plan: Added one new test for lambdas, otherwise this refactor maintains old functionality as evidenced by the tests that continue to work.

Reviewers: zasgar, michelle, nserrino, #engineering

Reviewed By: michelle, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D1454

GitOrigin-RevId: 8c7fa7262a50972b124ce7759f5b740ce6d1201c