Add RowBatch serialization/deserialization to protobuf.
Summary:
This is partially a refactor and partially new code.
For GRPC Sources/Sinks,we will need to be able to serialize/deserialize
Rowbatches to Protobuf. This is temporary until we make feather/flat buffers our
serialization format for RowBatches.

Previously in the table.cc code, there was serialization in the form of turning
RowBatches into arrow RecordBatches and then serializing those into RowBatches.
I moved that code and have it directly in the RowBatch class, because it makes sense
for the RowBatch class to know how (de)serialize itself.

This way there isn't duplication and it can be used by the GRPC sources/sinks.

Test Plan: added test for (de)serialization, existing tests.

Reviewers: michelle, zasgar, philkuz, #engineering

Reviewed By: michelle, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D1458

GitOrigin-RevId: cbf8186a93b8cc35f89138d106a64326b040f31e