ExecState needs ability to generate Kelvin service stubs (deprecate RowBatchQueue)
Summary:
The previous design for multinode queries involved the grpc sink exec node queuing
rowbatches, and that queue would be drained (outside of carnot) and the rowbatches sent to the remote
grpc service.

The new design actually has the grpc sink making the grpc calls directly. as a result it needs a way
to get a handle on the stub to the grpc service in both a test and a production context.

In the prod context we need the credentials that are declared in the agent. in the test context we
need the Mock of the Kelvin GRPC stub (autogenerated by the proto).

Test Plan: existing

Reviewers: zasgar, michelle, philkuz, #engineering

Reviewed By: michelle, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D1471

GitOrigin-RevId: 806853cb813d1159cc622739d727af57c73a98ba