Clean-up connections on unknown protocols
Summary:
Normally, connect/accept and close for unknown protocols are collected,
because the connection protocol may be discovered later.

This means a connection tracker is instantiated for a lot more connections than are actually useful.

Unfortunately, there was an issue where connection trackers for any connections that were not being
converted to a known protocol were not being cleaned up properly.

This fixes that.

Test Plan: New test added to socket_trace_connector_test.cc.

Reviewers: zasgar, yzhao, #engineering, chengruizhe

Reviewed By: yzhao, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D1494

GitOrigin-RevId: 151002be4688f3e9a37f2da715bc24a2a1c590fa