Maintain per-DataStream Inflater object for decompressing HTTP2 headers.
Summary:
This fixes the HTTP2 headers parse failure, by maintaining a per-stream Inflater
object that manages the decompression (deflation as used by nghttp2) state.

Previous code is broken, because it creates new inflater for each attempt of frame
stitching, which loses the previously encoded headers (HTTP2 spec requires
compression/deflation algorithm to use coded bytes to represent repeated texts),
therefore causes encoded headers cannot be decoded.

Add Inflater class, which wraps nghttp2_hd_inflater and implements RAII.

Opt to let DataStream own a std::unique_ptr<Inflater> to have resource management
automatic within the existing framework.

An alternative is to move Inflater outside of ConnectionTracker entirely,
and maintain a ConnectionTracker to Inflater (a pair of them actually, one for
requests and the other responses). This makes resource management cumbersome and error-prone,
because now the framework needs to manage an associated data structure in addition to
ConnectTracker.

Another slight tweak is to have Inflater owned by ConnectionTracker. Overall that seems equivalent
to the current approach. The current approach appears better in that conceptually Inflater is tied
to a unidirectional HTTP2 frames, so it is considered closer to DataStream than to
ConnectionTracker.

The test is rewritten substantially to reduce redundant. It was not necessary, but seems
useful and simplified the test.

Test Plan: Jenkins

Reviewers: oazizi, #engineering

Reviewed By: oazizi, #engineering

Subscribers: chengruizhe

Differential Revision: https://phab.corp.pixielabs.ai/D1511

GitOrigin-RevId: a4ee203a32ba256d24b8106e963cef94d5e7c8d8