Stirling: Stuck DataStream Detection
Summary:
Detect when a stream is stuck, meaning either:
 1) A lost event has occured.
 2) An unparseable sequence of events is blocking the stream.

Note this pertains only to the parsing of messages from raw events, not the matching of requests and responses.

Also include a placeholder recovery algorithm for lost events, that simply jumps over the lost events.

Test Plan: See connection_tracker_test

Reviewers: yzhao, zasgar, chengruizhe, #engineering

Reviewed By: yzhao, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D1508

GitOrigin-RevId: df77c7927659b2921af322ffa0f7113bc5405efe