Stirling: Refactor ProcessMessages into ConnectionTracker
Summary:
Code that is per connection should really be in ConnectionTracker.

SocketTraceConnector should only have to care about getting messages out of the ConnectionTracker,
and pushing them to the DataTable.

Note while this diff moves into a better direction, it is by no means complete. There is still organizational work to be done,
but going to do it a piece at a time, to minimize churn.

Test Plan:
No new tests, because this is just a refactor.
But in a future diff, it may make sense to transfer some of the tests from socket_trace_connector_test.cc
into connection_tracker_test.cc. That would also reduce the number of FRIEND_TESTs that we need,
because ConnectionTracker becomes a testable unit, and has the right public interface.

Reviewers: yzhao, chengruizhe, zasgar, #engineering

Reviewed By: yzhao, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D1521

GitOrigin-RevId: e87c18ff69b7eaa3ab83d4415735597b86c4ab04