Stirling/BPF: Don't add new_conn to map if aborting
Summary:
I noticed a large number of close connections into the perf buffer, without any other events.

After digging a bit, the root cause was that get_conn_info() in the read and write tracing
would create an entry in the conn_info_map, even if we would then abort and return 0.

To fix, this diff checks if any other events have been previously reported. If not, it won't send the close event either.

Test Plan: Manually tested to confirm that there are fewer close events getting to user space.

Reviewers: yzhao, zasgar, #engineering

Reviewed By: yzhao, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D1529

GitOrigin-RevId: 205be02fa5d6c784cfdbb439946480f88860f28a