PL-734 UI: Add DialogBox component to be used in signup/login/deploy page
Summary:
This just implements the simple dialog box, where the header contains the Pixie logo, and the contents can contain anything.
This will be used for the Claim your Site box, Log into your Company box, and the deploy pixie agent box.
Here's a screenshot of it in storybook:
{F17037}
You can also view it directly in storybook, but it's not super interesting.
The box will be centered on the page it's used in, but the component itself does not handle any of that logic

Test Plan: unit test

Reviewers: zasgar, yike, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D1541

GitOrigin-RevId: 5fd075547a4417722343cb8079ff693fde9fe284