(PL-825) Transfer proto generation to new IR::ToProto
Summary: proto generation was previously very messy and hard to debug, also seeminly misplaced. Moved out of compiler and into ir_nodes. Got rid of the old IRWalker scheme that was hard to work with and replaced it with using the Operator parent class, something that wasn't available when the first Serialization code for the plan was written. New code is easier to write and read.

Test Plan: No new functionality, just refactr. So far passes all tests.

Reviewers: zasgar, michelle, nserrino, #engineering

Reviewed By: michelle, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D1542

GitOrigin-RevId: 90b395cc4e07bacfe8f71c3a1483e5ef4bfa194e