Prevent outer joins on non-time table for time ordered join case
Summary:
We guarantee that we preserve time_ order in output tables generated by
operators. For joins, that time_ column comes from one of the two tables.
That means that every record in the output table much have column values set from
the table from which time_ comes. Otherwise time_ will not be set and the table
will have 0s instead of a legitimate time value.
It seems like it's ok to say that for the time-ordered case, only left and inner joins
are allowed (assuming the time_ column comes from the left table) and that approach
preserves the cleanliness of the time_ column. For the non-time-ordered case,
any type of join is allowed.

Test Plan: existing,added a unit test

Reviewers: philkuz, michelle, zasgar, #engineering

Reviewed By: michelle, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D1563

GitOrigin-RevId: 0d57140c8567ffed4293d3fc74abfd51cd424bd8