PL-809 Admin CLI: Implement "check" functionality for deploy cmd
Summary:
For the "deploy" cmd in the admin CLI, users should be able to pass the "--check" flag to run basic checks that ensures the k8s version and node kernel versions all meet the minimum requirements.
this diff implements that functionality, but does not actually implement the actual "deploy" functionality yet.

Test Plan: tried it out

Reviewers: zasgar, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D1535

GitOrigin-RevId: 6a764cf31a3e805cbcd690788e8248741f740a93