PL-734 Admin CLI: Deploy
Summary:
The user should be able to use the CLI to deploy the Pixie system.
Currently, I assume we will be packaging the vizier/vizier_dep yaml files that have already had "kustomize build" run on them. This is because any kubectl version under 1.14 does not have kustomize support. For now, I put these yamls inside a v1_yaml directory which we use to generate some bindata.
In the future, I think we'll want to write a script that generates these yaml files for each version of our system.

Test Plan: tried it out

Reviewers: zasgar, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D1609

GitOrigin-RevId: 79e17c6a3e2892581f4f5dcfa291fd920e803059