PL-734 UI: Auth0 Login UI
Summary:
we should have two separate pages, one for logging in and one for creating a new site. this updates the existing Auth.tsx to fit into the new directory structure we have.
also updated the styling for the auth0 lock widget.

Test Plan: tried manually... auth0 is hard to test in jest

Reviewers: zasgar, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D1641

GitOrigin-RevId: 2440735ed4d1f636f58cb9225e7b1bb087f2ef2e