Stirling: Enable stream recovery in connection tracker
Summary:
This diff is the last piece of the sequence of stream recovery diffs.

This piece uses the updated ParseMessages() with a sync_type to choose the type of recovery to perform.

Initially, we don't want to burn any cycles attempting recovery, but once a stream becomes stuck we go to progressively more aggressive forms of recovery.

Test Plan: A bunch of tests in connection_tracker_test.cc.

Reviewers: yzhao, zasgar, chengruizhe, #engineering

Reviewed By: yzhao, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D1553

GitOrigin-RevId: 775a70332c1a9325e20c3e86a8e2957a8baab581