Add parents to the dagnode proto.
Summary: The dag proto could not represent the order of parents accurately. As a result, Join nodes and operator nodes couldn't be properly stored as protos. This change adds a sorted parents field for the proto, which preserves the order of the parents in the DAG. Additionally added a ToProto, FromProto on the DAG node as this helps to integrate the new proto better.

Test Plan: add toproto/fromproto

Reviewers: zasgar, michelle, nserrino, #engineering

Reviewed By: michelle, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D1672

GitOrigin-RevId: 7db6a29d5c845d51bebc3277ef051f05ad86c58b