[Cleanup] Use templates to choose Send/Recv implementation ClientServerSystem.
Summary:
This diff is a stylistic clean-up, to use templates to select the send/recv implementation from TCPSocket to use.
This uses the power of templates to reduce the amount of duplicate code to support the different syscall variants.

Example of the new usage model:

```
ClientServerSystem system;
system.RunClientServer<&TCPSocket::Recv, &TCPSocket::Send>({{kHTTPRespMsg1, kHTTPRespMsg2}});
```

And if you want to use a different syscall function, you simply change the template arguments:

```
ClientServerSystem system;
system.RunClientServer<&TCPSocket::Read, &TCPSocket::Write>({{kHTTPRespMsg1, kHTTPRespMsg2}});
```

The appropriate member function is passed all the way through ClientServerSystem, and then it uses the
appropriate syscall to perform the send or recv.

The SendMsg/WriteV/ReadMsg/ReadV is also templatized in a similar manner. Note, however,
that since ReadMsg and ReadV do not have the exact same signature in TCPSocket, an extra set
of templated functions had to be created for this case.

Still, overall, the amount of virtually duplicate code has been reduced with this diff,
since the templates choose the correct function to use (and we don't have to have different versions of every
function).

Test Plan: This is the test - and it still works

Reviewers: yzhao, zasgar, #engineering

Reviewed By: yzhao, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D1674

GitOrigin-RevId: d64fdc2d918eed7f3cb2d0b091a1f3a4bf744235