Create service udfs to convert upids to services.
Summary: TSIA. Key things are that our current udfs return only services that are currently running. At the moment, we don't have good support to match a upid to a service that was running when the upid info was placed.

Test Plan: add tests that try to catch all the edge cases.

Reviewers: michelle, #engineering, zasgar

Reviewed By: #engineering, zasgar

Subscribers: zasgar

Differential Revision: https://phab.corp.pixielabs.ai/D1716

GitOrigin-RevId: fe7f57afa3889cc78ea858168a65af0289d6a6d4