Update plan proto and physical plan protos to support tablets.
Summary:
add plan proto update to support tablet as a tablet key. If table doesn't have a tablet, then the tablet key is not specified.

physical plan proto update to add table info,which includes table name, schema, tabletication_kye and the tablets that are available for the table.

Test Plan: n/a

Reviewers: zasgar, michelle, #engineering

Reviewed By: michelle, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D1673

GitOrigin-RevId: d8d3f0466602bad54a3071e0f44cc6b1259186d8