Remove columns vector from MemorySourceIR to remove unnecessary node dependency.
Summary: Wanted to simplify the copying of Memory Sources when we include tabletization. Realized the Column children of Memory source were unnecessary and added burden to copying so removed the dependency on them.

Test Plan: All old tests pass, had to renumber nodes because no need to create the columns anymore.

Reviewers: nserrino, michelle, zasgar, #engineering

Reviewed By: nserrino, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D1680

GitOrigin-RevId: 2adff3bdd5b1ee1dc60b50590afc326499f337c7