Admin CLI: Add image tag replacement to deploy cmd
Summary:
before, we were always just deploying the :latest image tag. we should take the version the user specifies in the cli and replace :latest with that image tag instead.
also update the jwt token so that its a valid string that can be used in the jwt_signing_key env variable.

Test Plan: tried it manually

Reviewers: zasgar, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D1728

GitOrigin-RevId: 3bf612fb7c3d0b08beb59a842ebfc21eabb28f8f