UI: Add timeout for vizier request
Summary:
We were starting to see weird behavior where the gql request would get stuck at "Loading"
This is because we had the old IP stored in localstorage (for example https://10.32.0.10:443). The UI would try to make a request to this IP, which would hang... for a very long time.
Added a timeout to the initial request so that it no longer hangs.

Test Plan: tried it manually

Reviewers: zasgar, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D1730

GitOrigin-RevId: d277c0b3c7e09e9e47604fc1a3df52ce9115d1b7