Imports protobuf as submodule
Summary:
The reason that it cannot be adapted as nghttp2, is that we need to change a piece of code
that is deep in the chain of DescriptorDatabase, DynamicMessage, DescriptorProto.

We'll need to replace DynamicMessage with our own version, and and reimplement DescriptorDatabase to
return that new class, that will be too much work and fragile to breakage as we'll duplicate a lot
of code.

Test Plan: Jenkins

Reviewers: #engineering, oazizi, zasgar

Reviewed By: #engineering, oazizi

Differential Revision: https://phab.corp.pixielabs.ai/D1742

GitOrigin-RevId: eb0debd44a97be5bb14598eece0f4cc0e0c75edf