Temporarily disable JWT valid user check
Summary:
Since we're currently using the same JWT for all of our service/cluster/user communications, we'll need to disable the valid user check for now until we implement the separate user/service/cluster jwts.
this is just so we can get to a working state until that work is done.

Test Plan: updated tests

Reviewers: zasgar, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D1740

GitOrigin-RevId: fe054089298b56954ad5b91fd81dc633a2ce9a1a