Add health checks for other deployments
Summary:
We add /healthz as a health check into all of our services so let's expose those in k8s too.
I couldn't figure out a way to remove this boilerplate via kustomize but lmk if anyone has knowledge or ideas of how to do so.

Test Plan: test it out

Reviewers: michelle, zasgar, #engineering

Reviewed By: michelle, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D1782

GitOrigin-RevId: c4bd31dc5d2e6a9d4572d1daa1e70d24e8676d29