New type of check
Summary:
Defining a new form of CHECK, extending those in the glog implementation.
The following table shows a comparison:

```
// MODE               DEBUG           PRODUCTION
// -------------------------------------------------------
// DCHECK             FATAL           nothing
// CHECK              FATAL           FATAL
// XCHECK             FATAL           ERROR
```

Another way to think of it is as a shortcut to LOG_IF(DFATAL, condition).

Also, open to a different name.

Test Plan: Added a test, but it needs to be run in both debug and non-debug modes to stress both cases. Jenkins should do this.

Reviewers: yzhao, zasgar, #engineering

Reviewed By: yzhao, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D1783

GitOrigin-RevId: 9ddd64a75ebf9b7e7aaf310490d9e69eaeddff10