Make bitsize for RSA for certs configurable
Summary:
This is the only applicable part now that D1688 is abandoned.
In the future if we have google http load balancers, they won't be able to use the default value
of 4096 bitsize for RSA, so let's make this configurable so we can use them later.

Test Plan: n/a

Reviewers: michelle, zasgar, #engineering

Reviewed By: michelle, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D1793

GitOrigin-RevId: 36f8ac2b1677ffa1310f8750874370facab74dd1